improving usability of process to load strategy signals from local files passing Context to Strategy when producing a Signal beginnings of work to enable loading data from CSV/TSV files
