+++ univocity-trader-core/src/main/java/com/univocity/trader/Exchange.java
@@ -47,11 +47,11 @@ public interface Exchange<T, C extends AccountConfiguration<C>> {
 
 	/**
 	 * Returns historical candles/ticks for a given symbol at the given time interval, for a given time period. Used to populate
-	 * the local database managed by {@link CandleRepository} for backtesting and {@link Strategy} development. If the live exchange
+	 * the local database managed by {@link DatabaseCandleRepository} for backtesting and {@link Strategy} development. If the live exchange
 	 * limits the number of candles returned at once, that limit should be respected, and the interval given by
 	 * {@code startTime} or {@code endTime} can be shortened internally.
 	 *
-	 * Method {@link CandleRepository#fillHistoryGaps(Exchange, String, Instant, TimeInterval)} will cycle through the gaps in history and fill them accordingly
+	 * Method {@link DatabaseCandleRepository#fillHistoryGaps(Exchange, String, Instant, TimeInterval)} will cycle through the gaps in history and fill them accordingly
 	 * by requesting more candles from the exchange as needed.
 	 *
 	 * @param symbol    the symbol whose latest price information should be returned from the exchange (e.g. BTCUSDT, MSFT, etc)
