+++ univocity-trader-core/src/main/java/com/univocity/trader/simulation/MarketSimulator.java
@@ -24,7 +24,7 @@ public abstract class MarketSimulator<C extends Configuration<C, A>, A extends A
 	private static final Logger log = LoggerFactory.getLogger(MarketSimulator.class);
 
 	private final Supplier<Exchange<?, A>> exchangeSupplier;
-	private CandleRepository candleRepository;
+	private DatabaseCandleRepository candleRepository;
 	private ExecutorService executor;
 
 	protected MarketSimulator(C configuration, Supplier<Exchange<?, A>> exchangeSupplier) {
@@ -37,8 +37,8 @@ public abstract class MarketSimulator<C extends Configuration<C, A>, A extends A
 		resetBalances();
 	}
 
-	protected CandleRepository createCandleRepository() {
-		return new CandleRepository(configure().database());
+	protected DatabaseCandleRepository createCandleRepository() {
+		return new DatabaseCandleRepository(configure().database());
 	}
 
 	@Override
@@ -258,7 +258,7 @@ public abstract class MarketSimulator<C extends Configuration<C, A>, A extends A
 	public final void backfillHistory() {
 		TreeSet<String> allSymbols = new TreeSet<>();
 		configuration.accounts().forEach(a -> allSymbols.addAll(a.symbolPairs().keySet()));
-		allSymbols.addAll(new CandleRepository(configure().database()).getKnownSymbols());
+		allSymbols.addAll(new DatabaseCandleRepository(configure().database()).getKnownSymbols());
 		backfillHistory(allSymbols);
 	}
 
@@ -274,7 +274,7 @@ public abstract class MarketSimulator<C extends Configuration<C, A>, A extends A
 	}
 
 	protected void backfillHistory(Exchange<?, A> exchange, Collection<String> symbols) {
-		CandleRepository candleRepository = new CandleRepository(configure().database());
+		DatabaseCandleRepository candleRepository = new DatabaseCandleRepository(configure().database());
 		final Instant start = simulation.backfillFrom().toInstant(ZoneOffset.UTC);
 		final Instant end = simulation.backfillTo().toInstant(ZoneOffset.UTC);
 		CandleHistoryBackfill backfill = new CandleHistoryBackfill(candleRepository);
@@ -292,7 +292,7 @@ public abstract class MarketSimulator<C extends Configuration<C, A>, A extends A
 		long startTime;
 	}
 
-	public final CandleRepository getCandleRepository() {
+	public final DatabaseCandleRepository getCandleRepository() {
 		if (candleRepository == null) {
 			candleRepository = createCandleRepository();
 		}
