+++ univocity-trader-core/src/main/java/com/univocity/trader/simulation/SignalReproducer.java
@@ -1,30 +1,28 @@
 package com.univocity.trader.simulation;
 
+import com.univocity.trader.account.*;
 import com.univocity.trader.candles.*;
 import com.univocity.trader.indicators.*;
 import com.univocity.trader.strategy.*;
 import org.slf4j.*;
 
-import java.util.*;
+import java.io.*;
 
 public class SignalReproducer implements Strategy {
 	private static final Logger log = LoggerFactory.getLogger(SignalReproducer.class);
 
-	private final Map<Candle, Signal> signals;
+	private final SignalRepository signalRepository;
 
-	public SignalReproducer(String symbol, SignalRepository repository) {
-		signals = repository.signalsFor(symbol);
-		if (signals == null) {
-			throw new IllegalArgumentException("No signals for " + symbol + " in repository");
-		}
+	public SignalReproducer(String symbol, Reader input) {
+		this(new SignalRepository(symbol, input));
+	}
+
+	public SignalReproducer(SignalRepository repository) {
+		this.signalRepository = repository;
 	}
 
 	@Override
-	public Signal getSignal(Candle candle) {
-		Signal out = signals.get(candle);
-		if (out != null) {
-			return out;
-		}
-		return Signal.NEUTRAL;
+	public Signal getSignal(Candle candle, Context context) {
+		return signalRepository.signalFor(context.symbol(), candle);
 	}
 }
